feat: Add demo code to handle spartan link
The demo embed extension demonstrates how to parse a link that is
can be recieved by a spartan user from a schedule dialog in an embed.
The link is receieved in the link param and is subsequently mapped to
its own route format.

Additionally:
1. Reduces the amount of data when requesting dashboard data.
2. Renames dashboard-next link to dashboard (this was changed in the
Looker UI.
3. Adds a command to start demo embed dev server from the root of the
repo (more commands will be added).
4. Updates gitignore to ignore direnv stuff.