feat: improve code splitting example (#105)
Updates code splitting example to reflect latest thinking. The
ExtensionProvider40 is now loaded separately from the application
reducing the initial bundle size from 1.1MB to 775kb. The thinking
is that the ExtensionProvider should be loaded as quickly as possible
so that it can communicate with the Looker host as soon as possible.